+++ src/Symfony/Component/Messenger/Bridge/Amqp/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 8.2
 ---
 
+ * Add the `prefetch_count` option to consume messages with a long lived consumer instead of fetching them one by one
  * Round delays up to two significant digits and add the `delay[granularity]` option to control that rounding
  * Implement the `KeepaliveReceiverInterface` to enable asynchronously notifying AMQP that the job is still being processed, in order to avoid timeouts
  * Pass the routing key to the serializer as `extra[routing_key]` when decoding